+++ common/src/main/java/com/faboslav/structurify/common/mixin/compat/repurposedstructures/RepurposedStructuresModifySpreadMixin.java
@@ -1,22 +1,23 @@
-package com.faboslav.structurify.common.mixin.structure.placement.compat;
-
-import org.spongepowered.asm.mixin.Mixin;
-import net.minecraft.world.level.levelgen.structure.placement.RandomSpreadStructurePlacement;
-
 //? if repurposed_structures {
+package com.faboslav.structurify.common.mixin.compat.repurposedstructures;
+
 import com.faboslav.structurify.common.api.StructurifyRandomSpreadStructurePlacement;
 import com.faboslav.structurify.common.util.RandomSpreadUtil;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import com.telepathicgrunt.repurposedstructures.world.structures.placements.AdvancedRandomSpread;
 import net.minecraft.core.Vec3i;
+import net.minecraft.world.level.levelgen.structure.placement.RandomSpreadStructurePlacement;
 import net.minecraft.world.level.levelgen.structure.placement.RandomSpreadType;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import java.util.Optional;
 
+@Pseudo
 @Mixin(value = AdvancedRandomSpread.class)
 public abstract class RepurposedStructuresModifySpreadMixin extends RandomSpreadStructurePlacement implements StructurifyRandomSpreadStructurePlacement
 {
@@ -79,10 +80,4 @@ public abstract class RepurposedStructuresModifySpreadMixin extends RandomSpread
 		return RandomSpreadUtil.getModifiedSeparation(this.structurify$getStructureSetId(), this.spacing(), originalSeparation);
 	}
 }
-//?} else {
-/*// This is just a placeholder mixin
-@Mixin(RandomSpreadStructurePlacement.class)
-public abstract class RepurposedStructuresModifySpreadMixin
-{
-}
-*///?}
+//?}