+++ forge/src/main/java/com/faboslav/structurify/forge/mixin/compat/structuregelapi/StructureGelApiGridStructurePlacement.java
@@ -1,4 +1,5 @@
-package com.faboslav.structurify.forge.mixin.compat;
+//? if structure_gel_api {
+package com.faboslav.structurify.forge.mixin.compat.structuregelapi;
 
 import com.faboslav.structurify.common.api.StructurifyRandomSpreadStructurePlacement;
 import com.faboslav.structurify.common.util.RandomSpreadUtil;
@@ -7,12 +8,14 @@ import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 
+@Pseudo
 @Mixin(value = GridStructurePlacement.class, remap = false)
-public abstract class StructureGelApiModifySpreadMixin implements StructurifyRandomSpreadStructurePlacement
+public abstract class StructureGelApiGridStructurePlacement implements StructurifyRandomSpreadStructurePlacement
 {
 	@Shadow
 	@Final
@@ -42,12 +45,31 @@ public abstract class StructureGelApiModifySpreadMixin implements StructurifyRan
 	public String structurify$getStructureSetId() {
 		return this.structurify$structureSetId;
 	}
+
 	public int structurify$getOriginalSpacing() {
 		return this.spacing;
 	}
 
 	public int structurify$getOriginalSeparation() {
 		return this.offset;
+	}
+
+	@ModifyReturnValue(
+		method = "salt",
+		at = @At("RETURN"),
+		remap = true
+	)
+	private int structurify$modifySalt(int originalSalt) {
+		return RandomSpreadUtil.getModifiedSalt(this.structurify$getStructureSetId(), originalSalt);
+	}
+
+	@ModifyReturnValue(
+		method = "frequency",
+		at = @At("RETURN"),
+		remap = true
+	)
+	private float structurify$modifyFrequency(float originalFrequency) {
+		return RandomSpreadUtil.getModifiedFrequency(this.structurify$getStructureSetId(), originalFrequency);
 	}
 
 	@ModifyReturnValue(
@@ -66,3 +88,4 @@ public abstract class StructureGelApiModifySpreadMixin implements StructurifyRan
 		return RandomSpreadUtil.getModifiedSeparation(this.structurify$getStructureSetId(), this.spacing(), originalOffset);
 	}
 }
+//?}